Separated GCP serial tests
